+++ Changelog.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Modify and standardized structure and method in `mariadb.js`. Completed on 2024-06-05
 - Alter rule strict checking in `onrequest` function at `reaction.js` and will force continue rules until the end event controller error happen. Completed on 2024-06-06
 - Minor change,update,remove unused code and comment in `sqlite3.js`. Completed on 2024-06-06
+- Simplify `query` function return data method in `sqlite3.js`. Completed on 2024-06-07
 
 ### Deprecated